requirements.txt
pandas
openpyxl
requests
streamlit
python-dotenv
apscheduler
supabase